Change according to review comments of #165
